Durability color and decimals, press "E"

+++ PROJ_P/Assets/Scripts/Attacks/Warrior/MeleeHack.cs
@@ -67,12 +67,12 @@ public class MeleeHack : PlayerAttack
     public void DecreaseDurability()
     {
         durability.DrainResource(durabilityDecrease);
-        durabilityText.text = ((durability.Value * 100)).ToString();
+        durabilityText.text = ((int)(durability.Value * 100)).ToString();
     }
     public void IncreaseDurability(float increase)
     {
         durability.IncreaseResource(increase);
-        durabilityText.text = ((durability.Value * 100)).ToString();
+        durabilityText.text = ((int)(durability.Value * 100)).ToString();
     }
     public void ResetSword()
     {
